XC-497: upgrade headers to directory
